+++ strategies/aggregateattestation/first/aggregateattestation_test.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Attestant Limited.
+// Copyright © 2020, 2021 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -38,6 +38,7 @@ func TestAggregateAttestation(t *testing.T) {
 			name: "Good",
 			params: []first.Parameter{
 				first.WithLogLevel(zerolog.Disabled),
+				first.WithTimeout(time.Second),
 				first.WithAggregateAttestationProviders(map[string]eth2client.AggregateAttestationProvider{
 					"good": mock.NewAggregateAttestationProvider(),
 				}),