+++ strategies/attestationdata/best/attestationdata.go
@@ -75,8 +75,9 @@ func (s *Service) AttestationData(ctx context.Context, slot phase0.Slot, committ
 			// Anyone not responded by now is considered errored.
 			errored = len(s.attestationDataProviders) - responded
 			log.Debug().Dur("elapsed", time.Since(started)).Int("responded", responded).Int("errored", errored).Msg("Timed out waiting for responses")
-		case <-errCh:
+		case err := <-errCh:
 			errored++
+			log.Debug().Dur("elapsed", time.Since(started)).Err(err).Msg("Responded with error")
 		case resp := <-respCh:
 			responded++
 			if bestAttestationData == nil || resp.score > bestScore {