+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 1.2.0 release candidate 1:
   - decrypt local accounts in parallel to reduce startup time
+  - add individual strategy timeouts for finer control of strategies
   - add 'advanced' scheduler, designed to be more robust with higher parallel job load
   - fetch wallet accounts from Dirk in parallel
   - fetch process-concurrency configuration value from most specific point in hierarchy